feat: (WIP) reorder chapters dnd
